Modified to long Ids Added base class for id entities Added Generic Id Entity Dao

+++ app/Seggu.Domain/ProducerCode.cs
@@ -5,8 +5,8 @@ namespace Seggu.Domain
     
     public partial class ProducerCode
     {
-        public int ProducerId { get; set; }
-        public int CompanyId { get; set; }
+        public long ProducerId { get; set; }
+        public long CompanyId { get; set; }
         public string Code { get; set; }
     
         public virtual Producer Producer { get; set; }
